Fix -Wshorten-64-to-32 warnings on 32 bit builds
git-svn-id: https://svn.osgeo.org/gdal/trunk@31647 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/vrt/vrtderivedrasterband.cpp
@@ -280,7 +280,7 @@ CPLErr VRTDerivedRasterBand::IRasterIO(GDALRWFlag eRWFlag,
 /* -------------------------------------------------------------------- */
     if ( nPixelSpace == typesize &&
          (!bNoDataValueSet || dfNoDataValue == 0) ) {
-        memset( pData, 0, nBufXSize * nBufYSize * nPixelSpace );
+        memset( pData, 0, static_cast<size_t>(nBufXSize * nBufYSize * nPixelSpace) );
     }
     else if ( !bEqualAreas || bNoDataValueSet )
     {
